バージョンナンバーを変更&ログインフォームをコンポーネントに分割
